1st version of corktile editing working, w/ drawbacks.
It still has not pretty layout.
It still doesn't update the modified in the `cork_corktiles` table.
It still isn't plugged to the dashboard.
It still does not treat properly location.
It still does not have inplace link to edit.
It still does not have caching optimizations.
It still does not have tests.
Among many other things...